pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask-login>=0.6.3",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "oauthlib>=3.2.2",
    "googletrans==3.1.0a0",
    "flask-socketio>=5.5.1",
    "eventlet>=0.39.0",
    "python-dotenv>=1.0.0",
    "pymongo[srv]>=4.6.1",
    "dnspython>=2.4.2",
    "flask-pymongo>=2.3.0",
]
